ml-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.3
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.1
aiofiles==23.2.1
prisma==0.15
